requirements.txt
# Core Django
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0

# Base de données
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# Fichiers statiques et média
whitenoise==6.6.0

# Stockage cloud (optionnel)
django-storages==1.14.2
boto3==1.35.0

# APIs et traitement
openai==1.101.0
requests==2.31.0

# PDF et documents
PyMuPDF==1.24.0
PyPDF2==3.0.1
reportlab==4.2.5
Pillow==11.3.0

# Scraping web
playwright==1.47.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilitaires
python-dotenv==1.1.1
asgiref==3.9.1
sqlparse==0.5.3

# Production
gunicorn==21.2.0
